+++ rocsolver/library/src/include/common_host_helpers.hpp
@@ -4,7 +4,6 @@
 
 #pragma once
 
-#include "common_ostream_helpers.hpp"
 #include <cassert>
 #include <cstdlib>
 #include <cstring>
@@ -79,54 +78,6 @@ template <typename T>
 T* const* cast2constPointer(T** array)
 {
     return array;
-}
-
-template <typename T, typename U, std::enable_if_t<!is_complex<T>, int> = 0>
-void print_device_matrix(const std::string name,
-                         const rocblas_int m,
-                         const rocblas_int n,
-                         U A,
-                         const rocblas_int lda)
-{
-    T hA[lda * n];
-    hipMemcpy(hA, A, sizeof(T) * lda * n, hipMemcpyDeviceToHost);
-
-    rocblas_cout << m << "-by-" << n << " matrix: " << name << std::endl;
-    for(int i = 0; i < m; i++)
-    {
-        rocblas_cout << "    ";
-        for(int j = 0; j < n; j++)
-        {
-            rocblas_cout << hA[j * lda + i];
-            if(j < n - 1)
-                rocblas_cout << ", ";
-        }
-        rocblas_cout << std::endl;
-    }
-}
-
-template <typename T, typename U, std::enable_if_t<is_complex<T>, int> = 0>
-void print_device_matrix(const std::string name,
-                         const rocblas_int m,
-                         const rocblas_int n,
-                         U A,
-                         const rocblas_int lda)
-{
-    T hA[lda * n];
-    hipMemcpy(hA, A, sizeof(T) * lda * n, hipMemcpyDeviceToHost);
-
-    rocblas_cout << m << "-by-" << n << " matrix: " << name << std::endl;
-    for(int i = 0; i < m; i++)
-    {
-        rocblas_cout << "    ";
-        for(int j = 0; j < n; j++)
-        {
-            rocblas_cout << '[' << hA[j * lda + i].real() << "+" << hA[j * lda + i].imag() << "i]";
-            if(j < n - 1)
-                rocblas_cout << ", ";
-        }
-        rocblas_cout << std::endl;
-    }
 }
 
 #ifdef ROCSOLVER_VERIFY_ASSUMPTIONS