+++ rocsolver/library/src/include/rocsolver_logger.hpp
@@ -9,6 +9,10 @@
 #include "rocsolver.h"
 #include <unordered_map>
 
+/***************************************************************************
+ * rocSOLVER logging macros
+ ***************************************************************************/
+
 #define ROCSOLVER_ENTER_TOP(name, ...)                                              \
     do                                                                              \
     {                                                                               \
@@ -53,25 +57,46 @@
         return _status;                                       \
     } while(0)
 
+/***************************************************************************
+ * The rocsolver_log_entry struct records function data for profile logging
+ * purposes.
+ ***************************************************************************/
 struct rocsolver_log_entry
 {
     std::string name;
     int calls;
     double time;
     double internal_time;
+
+    rocsolver_log_entry()
+        : calls(0)
+        , time(0)
+        , internal_time(0)
+    {
+    }
 };
 
+/***************************************************************************
+ * The rocsolver_logger class provides functions to be called upon entering
+ * or exiting a function that will output multi-level logging information.
+ ***************************************************************************/
 class rocsolver_logger
 {
 private:
+    // profile logging data keyed by function name
     std::unordered_map<std::string, rocsolver_log_entry> profile;
+    // function call stack keyed by handle
     std::unordered_map<rocblas_handle, std::vector<rocsolver_log_entry>> call_stack;
+    // the maximum depth at which nested function calls will appear in the log
     int max_levels;
+    // layer mode enum describing which logging facilities are enabled
     rocblas_layer_mode layer_mode;
 
+    // convert type T into char s, d, c, or z
     template <typename T>
     char get_precision();
 
+    // combines a function prefix and name into an std::string
     template <typename T>
     inline std::string get_func_name(const char* func_prefix, const char* func_name)
     {
@@ -82,26 +107,12 @@ private:
         return std::string(func_prefix) + '_' + func_name + "_template";
     }
 
+    // timing functions borrowed from rocblascommon/clients/include/utility.hpp
     double get_time_us();
     double get_time_us_sync(hipStream_t stream);
     double get_time_us_no_sync();
 
-    template <typename T1, typename T2, typename... Ts>
-    void log_arguments(rocsolver_ostream& os, const char* sep, T1 arg1, T2 arg2, Ts... args)
-    {
-        os << arg1 << ' ' << arg2;
-
-        if(sizeof...(Ts) > 0)
-        {
-            os << sep;
-            log_arguments(os, sep, args...);
-        }
-    }
-    void log_arguments(rocsolver_ostream& os, const char* sep)
-    {
-        // do nothing
-    }
-
+    // outputs bench logging
     template <typename T, typename... Ts>
     void log_bench(int level, const char* func_prefix, const char* func_name, Ts... args)
     {
@@ -109,10 +120,11 @@ private:
             rocblas_cout << "    ";
 
         rocblas_cout << "rocsolver-bench -f " << func_name << " -r " << get_precision<T>() << ' ';
-        log_arguments(rocblas_cout, " ", args...);
+        print_pairs(rocblas_cout, " ", args...);
         rocblas_cout << std::endl;
     }
 
+    // outputs trace logging
     template <typename T, typename... Ts>
     void log_trace(int level, const char* func_prefix, const char* func_name, Ts... args)
     {
@@ -120,10 +132,11 @@ private:
             rocblas_cout << "    ";
 
         rocblas_cout << get_template_name(func_prefix, func_name) << " (";
-        log_arguments(rocblas_cout, ", ", args...);
+        print_pairs(rocblas_cout, ", ", args...);
         rocblas_cout << ')' << std::endl;
     }
 
+    // populates profile logging data with information from call_stack
     template <typename T>
     void log_profile(rocblas_handle handle, int level, const char* func_prefix, const char* func_name)
     {
@@ -144,11 +157,13 @@ private:
     }
 
 public:
+    // returns true if logging facilities are enabled
     inline bool is_logging_enabled()
     {
         return layer_mode > 0;
     }
 
+    // logging function to be called upon entering a top-level (i.e. impl) function
     template <typename T, typename... Ts>
     void log_enter_top_level(rocblas_handle handle,
                              const char* func_prefix,
@@ -166,6 +181,7 @@ public:
             log_bench<T>(level, func_prefix, func_name, args...);
     }
 
+    // logging function to be called before exiting a top-level (i.e. impl) function
     template <typename T>
     void log_exit_top_level(rocblas_handle handle, const char* func_prefix, const char* func_name)
     {
@@ -180,6 +196,7 @@ public:
         call_stack.erase(handle);
     }
 
+    // logging function to be called upon entering a sub-level (i.e. template) function
     template <typename T, typename... Ts>
     void log_enter(rocblas_handle handle, const char* func_prefix, const char* func_name, Ts... args)
     {
@@ -194,6 +211,7 @@ public:
             log_trace<T>(level, func_prefix, func_name, args...);
     }
 
+    // logging function to be called before exiting a sub-level (i.e. template) function
     template <typename T>
     void log_exit(rocblas_handle handle, const char* func_prefix, const char* func_name)
     {