+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 Full documentation for rocSOLVER is available at [rocsolver.readthedocs.io](https://rocsolver.readthedocs.io/en/latest/).
 
+## [(Unreleased) rocSOLVER 3.12.0 for ROCm 4.2.0]
+### Added
+- Multi-level logging
+
+
 ## [(Unreleased) rocSOLVER 3.11.0 for ROCm 4.1.0]
 ### Added
 - Eigensolver routines for symmetric/hermitian matrices: